+++ llmfit-tui/src/main.rs
@@ -600,6 +600,7 @@ fn run_fit(
     perfect: bool,
     limit: Option<usize>,
     sort: SortColumn,
+    disp: &dyn display::DisplayMode,
     json: bool,
     memory_override: &Option<String>,
     context_limit: Option<u32>,
@@ -634,18 +635,14 @@ fn run_fit(
         fits.truncate(n);
     }
 
-    if json {
-        display::display_json_fits(&specs, &fits);
-    } else {
-        if hidden > 0 {
-            eprintln!(
-                "({} model{} hidden — incompatible backend)",
-                hidden,
-                if hidden == 1 { "" } else { "s" }
-            );
-        }
-        display::display_model_fits(&fits);
+    if hidden > 0 && !json {
+        eprintln!(
+            "({} model{} hidden — incompatible backend)",
+            hidden,
+            if hidden == 1 { "" } else { "s" }
+        );
     }
+    disp.display_model_fits(&specs, &fits);
 }
 
 fn fit_matches_filter(fit: &ModelFit, filter: FitArg) -> bool {
@@ -718,7 +715,7 @@ fn run_diff(
     fit_filter: FitArg,
     sort: SortColumn,
     limit: usize,
-    json: bool,
+    disp: &dyn display::DisplayMode,
     memory_override: &Option<String>,
     context_limit: Option<u32>,
 ) {
@@ -776,12 +773,7 @@ fn run_diff(
             fits.into_iter().take(limit).collect()
         };
 
-    if json {
-        display::display_json_diff_fits(&specs, &selected);
-    } else {
-        specs.display();
-        display::display_model_diff(&selected, sort.label());
-    }
+    disp.display_diff(&specs, &selected, sort.label());
 }
 
 fn run_tui(memory_override: &Option<String>, context_limit: Option<u32>) -> std::io::Result<()> {
@@ -868,6 +860,7 @@ fn run_recommend(
     min_fit: String,
     runtime_filter: String,
     capability: Option<String>,
+    disp: &dyn display::DisplayMode,
     json: bool,
     memory_override: &Option<String>,
     context_limit: Option<u32>,
@@ -956,14 +949,10 @@ fn run_recommend(
     fits = llmfit_core::fit::rank_models_by_fit(fits);
     fits.truncate(limit);
 
-    if json {
-        display::display_json_fits(&specs, &fits);
-    } else {
-        if !fits.is_empty() {
-            specs.display();
-        }
-        display::display_model_fits(&fits);
+    if !json && !fits.is_empty() {
+        specs.display();
     }
+    disp.display_model_fits(&specs, &fits);
 }
 
 fn run_download(
@@ -1278,7 +1267,7 @@ fn run_plan(
     context: u32,
     quant: Option<String>,
     target_tps: Option<f64>,
-    json: bool,
+    disp: &dyn display::DisplayMode,
     memory_override: &Option<String>,
 ) -> Result<(), String> {
     let db = ModelDatabase::new();
@@ -1292,12 +1281,7 @@ fn run_plan(
     };
     let plan = estimate_model_plan(model, &request, &specs)?;
 
-    if json {
-        display::display_json_plan(&plan);
-    } else {
-        specs.display();
-        display::display_model_plan(&plan);
-    }
+    disp.display_plan(&specs, &plan);
 
     Ok(())
 }
@@ -1306,29 +1290,20 @@ fn main() {
     let cli = Cli::parse();
     let context_limit = resolve_context_limit(cli.max_context);
 
+    // Build the display backend once, based on the --json flag.
+    let disp = display::new(cli.json);
+
     // If a subcommand is given, use classic CLI mode
     if let Some(command) = cli.command {
         match command {
             Commands::System => {
                 let specs = detect_specs(&cli.memory);
-                if cli.json {
-                    display::display_json_system(&specs);
-                } else {
-                    specs.display();
-                }
+                disp.display_system(&specs);
             }
 
             Commands::List => {
                 let db = ModelDatabase::new();
-                if cli.json {
-                    println!(
-                        "{}",
-                        serde_json::to_string_pretty(db.get_all_models())
-                            .expect("JSON serialization failed")
-                    );
-                } else {
-                    display::display_all_models(db.get_all_models());
-                }
+                disp.display_models(db.get_all_models());
             }
 
             Commands::Fit {
@@ -1340,6 +1315,7 @@ fn main() {
                     perfect,
                     limit,
                     sort.into(),
+                    disp.as_ref(),
                     cli.json,
                     &cli.memory,
                     context_limit,
@@ -1349,7 +1325,7 @@ fn main() {
             Commands::Search { query } => {
                 let db = ModelDatabase::new();
                 let results = db.find_model(&query);
-                display::display_search_results(&results, &query);
+                disp.display_search_results(&results, &query);
             }
 
             Commands::Info { model } => {
@@ -1366,11 +1342,7 @@ fn main() {
                 };
 
                 let fit = ModelFit::analyze_with_context_limit(&models[idx], &specs, context_limit);
-                if cli.json {
-                    display::display_json_fits(&specs, &[fit]);
-                } else {
-                    display::display_model_detail(&fit);
-                }
+                disp.display_model_detail(&specs, &fit);
             }
 
             Commands::Diff {
@@ -1386,7 +1358,7 @@ fn main() {
                     fit,
                     sort.into(),
                     limit,
-                    cli.json,
+                    disp.as_ref(),
                     &cli.memory,
                     context_limit,
                 );
@@ -1398,9 +1370,14 @@ fn main() {
                 quant,
                 target_tps,
             } => {
-                if let Err(err) =
-                    run_plan(&model, context, quant, target_tps, cli.json, &cli.memory)
-                {
+                if let Err(err) = run_plan(
+                    &model,
+                    context,
+                    quant,
+                    target_tps,
+                    disp.as_ref(),
+                    &cli.memory,
+                ) {
                     eprintln!("Error: {}", err);
                     std::process::exit(1);
                 }
@@ -1414,12 +1391,15 @@ fn main() {
                 capability,
                 json,
             } => {
+                // Recommend has its own --json flag; build a local display if needed.
+                let rec_disp = display::new(json);
                 run_recommend(
                     limit,
                     use_case,
                     min_fit,
                     runtime,
                     capability,
+                    rec_disp.as_ref(),
                     json,
                     &cli.memory,
                     context_limit,
@@ -1465,6 +1445,7 @@ fn main() {
             cli.perfect,
             cli.limit,
             cli.sort.into(),
+            disp.as_ref(),
             cli.json,
             &cli.memory,
             context_limit,
